a better hack: response builder returns void
According to the type system, response builder returns void. Which is what
we want when we're using tab completion. Void means there's nothing left
to add to the chain.

But when it runs in JS, it returns an object with the status code, headers,
and content. ts-node allows it because we turn off type checking at
runtime.

Might want to be more honest by not returning the response. Just call
respond(200).text("hello world") and don't return anything.