beginning of interceptors
registry.addInterceptor(url: string, callback: InterceptorCallback) ignores the url part and sets a global interceptor

The logic for normalizing a response has been moved from dispatcher to registry. That was necessary so that the respondTo() function passed to an interceptor returns a normalized response. The logic has not yet been removed from dispatcher.